Add MySQL support (#230)
With this commit, sqlc now supports two database engines: PostgreSQL and MySQL. The MySQL support contained in this change is experimental. It still fails to parse many valid queries, and the code generated for some queries is incorrect.

Also, the interface between parsing, codegen, and database catalog is a bit of a mess. The goal of merging this support in now is to make progress towards 1.0

Co-authored-by: Charles Moog <moogcharlie@gmail.com>