internal/dinosql: strip leading "go-" or trailing "-go" from import (#262)
If I try to import from a package named "go-x" or "x-go" it will
generate a syntax error. We should do the right thing and determine
what the package name is and set it to that, but in lieu of that,
a good guess for the import name is just stripping that prefix or
suffix.

Updates #261.